Added basic 'progression'
Two new functions in game master, levelWin and nextLevel, as well as a
new UI canvas for the win screen which is currently a placeholder.
levelWin should be called when the level is won, at which point the win
screen is displayed with a image and a next level button.  Clicking the
next level button triggers nextLevel, which hides the win screen and
just calls reset, but will need to be modified.