New Collision System
Got all the new raycast stuff down but,
Traps are in an interesting situation.

Trap are detected first before walls, which makes putting the return
statement in the for loop bad leading to a few deaths by clipping
through wall.  If the if statement with the death flag has another
argument, it affects character enemy interactions. Without the argument,
it'll even if the character goes on a different nontrap tile.